fix(test): sort results in example test
by sorting the results that we would like to get from redis
we make sure the exampele test passes.

p.s. we are sorting by price but there are items with identical prices
the data itself may need to change a bit. Let's see if we experience any
flakiness with this setup.