ReportResult should be given the result if it is a success
according to the interface, successes will be given to ReportResult as
err == nil.

In releaseConn it's given the err regardless of nil or non nil. But with
getConn, it is only given err if err != nil.

I don't think this should be the behaviour because right before getConn
is called, there's an Allow() call. I can't speak for everyone's
use-case, but I'm setting Allow() to return true, and looking for a
success call next - and this usually works - but sometimes I get stuck,
and I think it's because in the case where Allow() returns true for a
_getConn, if this succeeds, ReportResult is not given a success.